+++ R/fit_members.R
@@ -164,7 +164,7 @@ fit_member <- function(name, wflows, members_map, train_dat) {
     
     new_member <- 
       tune::finalize_workflow(member_wf, member_metrics[,member_params]) %>%
-      generics::fit(data = train_dat)
+      parsnip::fit(data = train_dat)
   } else {
     member_model <-
       members_map %>%
@@ -173,7 +173,7 @@ fit_member <- function(name, wflows, members_map, train_dat) {
       dplyr::pull()
     
     new_member <-
-      generics::fit(wflows[[member_model[1]]], data = train_dat)
+      parsnip::fit(wflows[[member_model[1]]], data = train_dat)
   }
   
   new_member