+++ tests/testthat/test_add_candidates.R
@@ -1,5 +1,3 @@
-context("add_candidates")
-
 if ((!on_cran()) || interactive()) {
   if (on_github()) {
     load(paste0(Sys.getenv("GITHUB_WORKSPACE"), "/tests/testthat/helper_data.Rda"))
@@ -18,8 +16,6 @@ skip_if_not_installed("nnet")
 library(nnet)
 
 test_that("stack can add candidates (regression)", {
-  skip("still some inconsistencies with stored objects")
-  
   expect_equal(
     st_0 %>% add_candidates(reg_res_svm),
     st_reg_1
@@ -36,8 +32,6 @@ test_that("stack can add candidates (regression)", {
 })
 
 test_that("stack can add candidates (multinomial classification)", {
-  skip("still some inconsistencies with stored objects")
-  
   expect_equal(
     st_0 %>% add_candidates(class_res_rf),
     st_class_1
@@ -53,8 +47,6 @@ test_that("stack can add candidates (multinomial classification)", {
 })
 
 test_that("stack can add candidates (two-way classification)", {
-  skip("still some inconsistencies with stored objects")
-  
   expect_equal(
     st_0 %>% add_candidates(log_res_rf),
     st_log_1
@@ -155,6 +147,7 @@ test_that("add_candidates errors informatively with bad arguments", {
     "not generated with the appropriate control settings"
   )
   
+  suppressWarnings(
   reg_res_lr_bad2 <- 
     tune::fit_resamples(
       object = reg_wf_lr,
@@ -165,6 +158,7 @@ test_that("add_candidates errors informatively with bad arguments", {
         save_workflow = TRUE
       )
     )
+  )
   
   expect_error(
     stacks() %>% add_candidates(reg_res_lr_bad2),
@@ -206,6 +200,34 @@ test_that("add_candidates errors informatively with bad arguments", {
   expect_error(
     stacks() %>% add_candidates(log_res),
     "only metrics that rely on hard class predictions"
+  )
+  
+  # warn when stacking may fail due to tuning failure
+  set.seed(7898)
+  data_folds <- rsample::vfold_cv(mtcars, v = 2)
+  
+  rec <- recipes::recipe(mpg ~ ., data = mtcars) %>%
+    recipes::step_bs(disp, deg_free = tune())
+  
+  model <- parsnip::linear_reg(mode = "regression", penalty = tune::tune()) %>%
+    parsnip::set_engine("glmnet")
+  
+  cars_grid_1 <- tibble::tibble(deg_free = 10L, penalty = -1)
+  
+  res_w_notes <- tune::tune_grid(
+    preprocessor = rec, 
+    object = model,
+    resamples = data_folds, 
+    grid = cars_grid_1, 
+    control = tune::control_grid(extract = function(x) {1}, save_pred = TRUE, save_workflow = TRUE)
+  )
+  
+  expect_error(
+    expect_warning(
+      stacks() %>%
+        add_candidates(res_2),
+      "argument \\`res_2\\` generated notes"
+    )
   )
 })