+++ NEWS.md
@@ -1,6 +1,4 @@
-# stacks 0.2.2.9001 (development version)
-
-To be released as v0.2.3.
+# stacks 0.2.3
 
 * Addressed deprecation warning in `add_candidates` (#99).
 * Improved clarity of warnings/errors related to failed hyperparameter